Call floor fixtures the same thing on the cursors as in the index
The entrance overview named its fixture cursors from the raw placement
id, so the store hung tickets reading "PV-DRAPE-TABLE-FRONT" and
"PROMO-STAND-BACK" over the floor -- engineering identifiers in the
biggest type on screen -- while the jump index called those same two
fixtures "PREVIOUSLY VIEWED" and "SUMMER SMASH HITS". The index also
numbered its repeats; the overview showed four cursors all reading
"VIDEO GAMES" with no way to tell them apart.

The index's naming was the right one, so it moves to a shared module
(fixture-labels.ts) and the overview uses it: a fixture's live identity
string -- a promo stand's current campaign topper, a bin's name -- with
the placement option and then the id as fallbacks, plus the same
repeat-qualification pass, applied here in left-to-right order because
that is the order the cursor ring steps them in.
